+++ backend/src/main/java/com/ssafy/cheertogether/league/dto/LeagueResponse.java
@@ -5,8 +5,6 @@ import java.util.List;
 import com.ssafy.cheertogether.league.domain.League;
 import com.ssafy.cheertogether.team.domain.Team;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -17,10 +15,10 @@ public class LeagueResponse {
 	private String name;
 	private String hanName;
 	private String logo;
-	private int apiId;
+	private Long apiId;
 	private List<Team> teamList;
 
-	public LeagueResponse (League league) {
+	public LeagueResponse(League league) {
 		this.name = league.getName();
 		this.hanName = league.getHanName();
 		this.logo = league.getLogo();
